+++ extract_msg_tests/__init__.py
@@ -1,6 +1,7 @@
 __all__ = [
     'AttachmentTests',
     'CommandLineTests',
+    'MessageTests',
     'OleWriterEditingTests',
     'OleWriterExportTests',
     'PropTests',
@@ -10,6 +11,7 @@ __all__ = [
 
 from .attachment_tests import AttachmentTests
 from .cmd_line_tests import CommandLineTests
+from .message_tests import MessageTests
 from .ole_writer_tests import OleWriterEditingTests, OleWriterExportTests
 from .prop_tests import PropTests
 from .util_tests import UtilTests